Change mod_rewrite check to use per-request note
Changes the logic to re-set the HTTPS environment variable when mod_rewrite causes a request to be reprocessed. Previously this used a connection-level note to track. This meant that subsequent HTTP requests on the same connection have the note set, and so exit the handler without being properly processed. This change sets a request-level note to be used by the mod_rewrite check, but leaves the connection-level note to allow ssl_is_https to work.